Finished up login and signup

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,12 +1,22 @@
+import { useAuthContext } from "../hooks/useAuthContext";
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
+
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch(
       `http://localhost:4000/api/workouts/${workout._id}`,
       {
         method: "DELETE",
+        headers: {
+          Authorization: `Bearer ${user.token}`,
+        },
       }
     );
